refactor: 주문 컨텍스트에서만 사용되는 키친라이더스 모듈을 order 패키지로 이동
